Separate category specific UI from the packager
