+++ src/ui_fsmenu/addons_packager.cc
@@ -22,6 +22,7 @@
 #include <memory>
 
 #include "base/i18n.h"
+#include "base/warning.h"
 #include "graphic/image_cache.h"
 #include "graphic/style_manager.h"
 #include "io/fileread.h"
@@ -30,7 +31,6 @@
 #include "io/profile.h"
 #include "logic/addon.h"
 #include "logic/filesystem_constants.h"
-#include "map_io/map_elemental_packet.h"
 #include "ui_basic/messagebox.h"
 #include "ui_basic/text_prompt.h"
 #include "ui_fsmenu/addons.h"
@@ -61,14 +61,7 @@ AddOnsPackager::AddOnsPackager(MainMenu& parent, AddOnsCtrl& ctrl)
         &box_right_subbox_header_hbox_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Vertical),
      box_right_subbox_header_box_right_(
         &box_right_subbox_header_hbox_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Vertical),
-     box_right_subbox_maps_dirstruct_hbox_(
-        &box_right_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Horizontal),
-     box_right_subbox_dirstruct_(
-        &box_right_subbox_maps_dirstruct_hbox_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Vertical),
-     box_right_subbox_maps_list_(
-        &box_right_subbox_maps_dirstruct_hbox_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Vertical),
-     box_right_buttonsbox_(
-        &box_right_subbox_maps_dirstruct_hbox_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Vertical),
+	 box_right_addon_specific_(&box_right_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Horizontal),
      box_right_bottombox_(&box_right_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Horizontal),
      name_(&box_right_subbox_header_box_right_, 0, 0, 100, UI::PanelStyle::kFsMenu),
      author_(&box_right_subbox_header_box_right_, 0, 0, 100, UI::PanelStyle::kFsMenu),
@@ -117,36 +110,7 @@ AddOnsPackager::AddOnsPackager(MainMenu& parent, AddOnsCtrl& ctrl)
          kButtonSize,
          UI::ButtonStyle::kFsMenuPrimary,
          _("OK")),
-     map_add_(&box_right_buttonsbox_,
-              "map_add",
-              0,
-              0,
-              kButtonSize,
-              kButtonSize,
-              UI::ButtonStyle::kFsMenuSecondary,
-              _("+"),
-              _("Add selected map")),
-     map_add_dir_(&box_right_buttonsbox_,
-                  "map_add_dir",
-                  0,
-                  0,
-                  kButtonSize,
-                  kButtonSize,
-                  UI::ButtonStyle::kFsMenuSecondary,
-                  _("*"),
-                  _("Create subdirectory")),
-     map_delete_(&box_right_buttonsbox_,
-                 "map_delete",
-                 0,
-                 0,
-                 kButtonSize,
-                 kButtonSize,
-                 UI::ButtonStyle::kFsMenuSecondary,
-                 _("–"),
-                 _("Remove selected map or directory")),
      addons_(&box_left_, 0, 0, 250, 0, UI::PanelStyle::kFsMenu),
-     dirstruct_(&box_right_subbox_dirstruct_, 0, 0, 200, 0, UI::PanelStyle::kFsMenu),
-     my_maps_(&box_right_subbox_maps_list_, 0, 0, 100, 0, UI::PanelStyle::kFsMenu),
      update_in_progress_(false) {
 
 	box_left_buttons_.add_inf_space();
@@ -191,37 +155,6 @@ AddOnsPackager::AddOnsPackager(MainMenu& parent, AddOnsCtrl& ctrl)
 	box_right_subbox_header_hbox_.add(
 	   &box_right_subbox_header_box_right_, UI::Box::Resizing::kExpandBoth);
 
-	box_right_buttonsbox_.add_inf_space();
-	box_right_buttonsbox_.add(&map_add_);
-	box_right_buttonsbox_.add_space(kSpacing);
-	box_right_buttonsbox_.add(&map_delete_);
-	box_right_buttonsbox_.add_inf_space();
-	box_right_buttonsbox_.add(&map_add_dir_);
-	box_right_buttonsbox_.add_inf_space();
-
-	box_right_subbox_dirstruct_.add(
-	   new UI::Textarea(&box_right_subbox_dirstruct_, UI::PanelStyle::kFsMenu,
-	                    UI::FontStyle::kFsGameSetupHeadings, _("Directory Tree"),
-	                    UI::Align::kCenter),
-	   UI::Box::Resizing::kFullSize);
-	box_right_subbox_dirstruct_.add_space(kSpacing);
-	box_right_subbox_dirstruct_.add(&dirstruct_, UI::Box::Resizing::kExpandBoth);
-
-	box_right_subbox_maps_list_.add(
-	   new UI::Textarea(&box_right_subbox_maps_list_, UI::PanelStyle::kFsMenu,
-	                    UI::FontStyle::kFsGameSetupHeadings, _("My Maps"), UI::Align::kCenter),
-	   UI::Box::Resizing::kFullSize);
-	box_right_subbox_maps_list_.add_space(kSpacing);
-	box_right_subbox_maps_list_.add(&my_maps_, UI::Box::Resizing::kExpandBoth);
-
-	box_right_subbox_maps_dirstruct_hbox_.add(
-	   &box_right_subbox_dirstruct_, UI::Box::Resizing::kExpandBoth);
-	box_right_subbox_maps_dirstruct_hbox_.add_space(kSpacing);
-	box_right_subbox_maps_dirstruct_hbox_.add(&box_right_buttonsbox_, UI::Box::Resizing::kFullSize);
-	box_right_subbox_maps_dirstruct_hbox_.add_space(kSpacing);
-	box_right_subbox_maps_dirstruct_hbox_.add(
-	   &box_right_subbox_maps_list_, UI::Box::Resizing::kExpandBoth);
-
 	box_right_bottombox_.add(&discard_changes_, UI::Box::Resizing::kExpandBoth);
 	box_right_bottombox_.add_space(kSpacing);
 	box_right_bottombox_.add(&write_changes_, UI::Box::Resizing::kExpandBoth);
@@ -230,7 +163,7 @@ AddOnsPackager::AddOnsPackager(MainMenu& parent, AddOnsCtrl& ctrl)
 
 	box_right_.add(&box_right_subbox_header_hbox_, UI::Box::Resizing::kFullSize);
 	box_right_.add_space(kButtonSize);
-	box_right_.add(&box_right_subbox_maps_dirstruct_hbox_, UI::Box::Resizing::kExpandBoth);
+	box_right_.add(&box_right_addon_specific_, UI::Box::Resizing::kExpandBoth);
 	box_right_.add_space(kSpacing);
 	box_right_.add(&box_right_bottombox_, UI::Box::Resizing::kFullSize);
 
@@ -242,49 +175,16 @@ AddOnsPackager::AddOnsPackager(MainMenu& parent, AddOnsCtrl& ctrl)
 	main_box_.add_space(kSpacing);
 	main_box_.add(&box_right_, UI::Box::Resizing::kExpandBoth);
 
-	{
-		std::vector<MainMenu::MapEntry> v;
-		MainMenu::find_maps("maps/My_Maps", v);
-		for (const MainMenu::MapEntry& entry : v) {
-			my_maps_.add(
-			   entry.first.localized_name, entry.first.filename, nullptr, false,
-			   (boost::format("%s<br>%s<br>%s<br>%s<br>%s") %
-			    g_style_manager->font_style(UI::FontStyle::kFsTooltipHeader)
-			       .as_font_tag(entry.first.filename) %
-			    (boost::format(_("Name: %s")) %
-			     g_style_manager->font_style(UI::FontStyle::kFsMenuInfoPanelParagraph)
-			        .as_font_tag(entry.first.localized_name))
-			       .str() %
-			    (boost::format(_("Size: %s")) %
-			     g_style_manager->font_style(UI::FontStyle::kFsMenuInfoPanelParagraph)
-			        .as_font_tag(
-			           (boost::format(_("%1$u×%2$u")) % entry.first.width % entry.first.height).str()))
-			       .str() %
-			    (boost::format(_("Players: %s")) %
-			     g_style_manager->font_style(UI::FontStyle::kFsMenuInfoPanelParagraph)
-			        .as_font_tag(std::to_string(entry.first.nrplayers)))
-			       .str() %
-			    (boost::format(_("Description: %s")) %
-			     g_style_manager->font_style(UI::FontStyle::kFsMenuInfoPanelParagraph)
-			        .as_font_tag(entry.first.description))
-			       .str())
-			      .str());
-		}
+	addon_boxes_[AddOns::AddOnCategory::kMaps] = std::unique_ptr<MapsAddOnsPackagerBox>(
+				new MapsAddOnsPackagerBox(parent, &box_right_addon_specific_, UI::PanelStyle::kFsMenu, 0, 0, UI::Box::Horizontal));
+
+	for (auto&& pair : addon_boxes_) {
+		pair.second->set_modified_callback([this] { current_addon_edited(); });
 	}
 
-	map_add_.set_enabled(false);
-	my_maps_.selected.connect([this](uint32_t) { map_add_.set_enabled(true); });
-	map_delete_.set_enabled(false);
-	dirstruct_.selected.connect([this](uint32_t i) { map_delete_.set_enabled(i > 0); });
 	addons_.selected.connect([this](uint32_t) { addon_selected(); });
 	addon_new_.sigclicked.connect([this]() { clicked_new_addon(); });
 	addon_delete_.sigclicked.connect([this]() { clicked_delete_addon(); });
-	map_add_.sigclicked.connect(
-	   [this]() { clicked_add_or_delete_map_or_dir(ModifyAction::kAddMap); });
-	map_add_dir_.sigclicked.connect(
-	   [this]() { clicked_add_or_delete_map_or_dir(ModifyAction::kAddDir); });
-	map_delete_.sigclicked.connect(
-	   [this]() { clicked_add_or_delete_map_or_dir(ModifyAction::kDeleteMapOrDir); });
 	discard_changes_.sigclicked.connect([this]() { clicked_discard_changes(); });
 	write_changes_.sigclicked.connect([this]() { clicked_write_changes(); });
 	ok_.sigclicked.connect([this]() { die(); });
@@ -298,13 +198,6 @@ AddOnsPackager::AddOnsPackager(MainMenu& parent, AddOnsCtrl& ctrl)
 
 	layout();
 	center_to_parent();
-}
-
-AddOnsPackager::~AddOnsPackager()
-{
-	for (auto pair : mutable_addons_) {
-		delete pair.second;
-	}
 }
 
 bool AddOnsPackager::handle_key(const bool down, const SDL_Keysym code) {
@@ -323,14 +216,11 @@ void AddOnsPackager::layout() {
 }
 
 void AddOnsPackager::initialize_mutable_addons() {
-	for (auto pair : mutable_addons_) {
-		delete pair.second;
-	}
 	mutable_addons_.clear();
 	addons_with_changes_.clear();
 
 	for (const AddOns::AddOnState& a : AddOns::g_addons) {
-		mutable_addons_[a.first.internal_name] = AddOns::create_mutable_addon(a.first);
+		mutable_addons_[a.first.internal_name] = AddOns::Addon::create_mutable_addon(a.first);
 	}
 
 	rebuild_addon_list("");
@@ -350,7 +240,7 @@ void AddOnsPackager::rebuild_addon_list(const std::string& select) {
 }
 
 inline AddOns::Addon* AddOnsPackager::get_selected() {
-	return addons_.has_selection() ? mutable_addons_.at(addons_.get_selected()) : nullptr;
+	return addons_.has_selection() ? mutable_addons_.at(addons_.get_selected()).get() : nullptr;
 }
 
 void AddOnsPackager::addon_selected() {
@@ -358,7 +248,10 @@ void AddOnsPackager::addon_selected() {
 
 	addon_delete_.set_enabled(selected);
 	box_right_subbox_header_hbox_.set_visible(false);
-	box_right_subbox_maps_dirstruct_hbox_.set_visible(false);
+	for (auto&& pair : addon_boxes_) {
+		pair.second->set_visible(false);
+	}
+	box_right_addon_specific_.clear();
 
 	if (!selected) {
 		return;
@@ -373,56 +266,18 @@ void AddOnsPackager::addon_selected() {
 	version_.set_text(selected->get_version());
 	update_in_progress_ = false;
 
-	if (selected->get_category() != AddOns::AddOnCategory::kMaps) {
+	if (addon_boxes_.count(selected->get_category()) == 0) {
 		return;
 	}
 
-	box_right_subbox_maps_dirstruct_hbox_.set_visible(true);
-	rebuild_dirstruct(selected);
+	AddOnsPackagerBox* box = addon_boxes_[selected->get_category()].get();
+	box_right_addon_specific_.add(box, UI::Box::Resizing::kExpandBoth);
+	box->set_visible(true);
+	box->load_addon(selected);
+	layout();
 }
 
-void AddOnsPackager::rebuild_dirstruct(AddOns::Addon* a, const std::vector<std::string>& select) {
-	const std::string path = kAddOnDir + FileSystem::file_separator() + a->get_internal_name();
 
-	dirstruct_.clear();
-	dirstruct_to_tree_map_.clear();
-
-	std::vector<std::string> toplevel_entry;
-	dirstruct_to_tree_map_.push_back(toplevel_entry);
-
-	dirstruct_.add(a->get_internal_name(), path, g_image_cache->get("images/ui_basic/ls_wlscenario.png"),
-	               toplevel_entry == select, _("Top-level directory"));
-	assert(a->get_category() == AddOns::AddOnCategory::kMaps);
-	do_recursively_rebuild_dirstruct(dynamic_cast<AddOns::MapsAddon*>(a)->get_tree(), 1, path, toplevel_entry, select);
-}
-
-void AddOnsPackager::do_recursively_rebuild_dirstruct(const AddOns::MapsAddon::DirectoryTree* tree,
-                                                      const unsigned level,
-                                                      const std::string& path,
-                                                      const std::vector<std::string>& map_path,
-                                                      const std::vector<std::string>& select) {
-	for (const auto& pair : tree->subdirectories) {
-		std::vector<std::string> entry = map_path;
-		entry.push_back(pair.first);
-		dirstruct_to_tree_map_.push_back(entry);
-
-		const std::string subdir = path + FileSystem::file_separator() + pair.first;
-		dirstruct_.add(pair.first, subdir, g_image_cache->get("images/ui_basic/ls_dir.png"),
-		               entry == select, pair.first, "", level);
-
-		do_recursively_rebuild_dirstruct(&pair.second, level + 1, subdir, entry, select);
-	}
-
-	for (const auto& pair : tree->maps) {
-		std::vector<std::string> entry = map_path;
-		entry.push_back(pair.first);
-		dirstruct_to_tree_map_.push_back(entry);
-
-		dirstruct_.add(pair.first, path + FileSystem::file_separator() + pair.first,
-		               g_image_cache->get("images/ui_basic/ls_wlmap.png"), entry == select,
-		               pair.first, "", level);
-	}
-}
 
 void AddOnsPackager::current_addon_edited() {
 	if (update_in_progress_) {
@@ -430,7 +285,7 @@ void AddOnsPackager::current_addon_edited() {
 	}
 
 	const std::string& sel = addons_.get_selected();
-	AddOns::Addon* m = mutable_addons_.at(sel);
+	AddOns::Addon* m = mutable_addons_.at(sel).get();
 
 	m->update_info(name_.text(), author_.text(), descr_.get_text(), version_.text());
 
@@ -499,7 +354,7 @@ void AddOnsPackager::clicked_new_addon() {
 			{},
 			{}
 		};
-		mutable_addons_[name] = AddOns::create_mutable_addon(a);
+		mutable_addons_[name] = AddOns::Addon::create_mutable_addon(a);
 		addons_with_changes_[name] = false;
 		check_for_unsaved_changes();
 		rebuild_addon_list(name);
@@ -528,102 +383,6 @@ void AddOnsPackager::clicked_delete_addon() {
 	mutable_addons_.erase(it);
 	check_for_unsaved_changes();
 	rebuild_addon_list("");
-}
-
-void AddOnsPackager::clicked_add_or_delete_map_or_dir(const ModifyAction action) {
-	AddOns::Addon* m = get_selected();
-	assert(m);
-	assert(m->get_category() == AddOns::AddOnCategory::kMaps);
-
-	AddOns::MapsAddon::DirectoryTree* tree = dynamic_cast<AddOns::MapsAddon*>(m)->get_tree();
-	AddOns::MapsAddon::DirectoryTree* tree_parent = nullptr;
-	std::vector<std::string> select = dirstruct_to_tree_map_[dirstruct_.selection_index()];
-
-	std::string selected_map;
-	if (!select.empty() && select.back().size() >= kWidelandsMapExtension.size() &&
-	    select.back().compare(select.back().size() - kWidelandsMapExtension.size(),
-	                          kWidelandsMapExtension.size(), kWidelandsMapExtension) == 0) {
-		// Last entry is a map – pop it, we care only about directories here
-		selected_map = select.back();
-		select.pop_back();
-	}
-	for (const std::string& s : select) {
-		tree_parent = tree;
-		tree = &tree->subdirectories.at(s);
-	}
-
-	switch (action) {
-	case ModifyAction::kAddMap: {
-		const std::string& map = my_maps_.get_selected();
-		const std::string filename = FileSystem::fs_filename(map.c_str());
-		tree->maps[filename] = map;
-		select.push_back(filename);
-		break;
-	}
-	case ModifyAction::kAddDir: {
-		UI::TextPrompt n(main_menu_, UI::WindowStyle::kFsMenu, _("New Directory"),
-		                 _("Enter the name for the new directory."));
-		for (;;) {
-			if (n.run<UI::Panel::Returncodes>() != UI::Panel::Returncodes::kOk) {
-				return;
-			}
-
-			std::string name = n.text();
-
-			if (name.empty() || !FileSystem::is_legal_filename(name) ||
-			    (name.size() >= kWidelandsMapExtension.size() &&
-			     name.compare(name.size() - kWidelandsMapExtension.size(),
-			                  kWidelandsMapExtension.size(), kWidelandsMapExtension) == 0)) {
-				main_menu_.show_messagebox(
-				   _("Invalid Name"), _("This name is invalid, please choose a different name."));
-				continue;
-			}
-
-			if (tree->subdirectories.find(name) != tree->subdirectories.end()) {
-				main_menu_.show_messagebox(
-				   _("Invalid Name"), _("A directory with this name already exists."));
-				continue;
-			}
-
-			tree->subdirectories[name] = AddOns::MapsAddon::DirectoryTree();
-			select.push_back(name);
-			break;
-		}
-		break;
-	}
-	case ModifyAction::kDeleteMapOrDir: {
-		assert(!selected_map.empty() || !select.empty());
-
-		UI::WLMessageBox mbox(
-		   get_parent(), UI::WindowStyle::kFsMenu, _("Delete"),
-		   selected_map.empty() ?
-		      (boost::format(
-		          _("Do you really want to delete the directory ‘%s’ and all its contents?")) %
-		       select.back())
-		         .str() :
-		      (boost::format(_("Do you really want to delete the map ‘%s’?")) % selected_map).str(),
-		   UI::WLMessageBox::MBoxType::kOkCancel, UI::Align::kLeft);
-		if (mbox.run<UI::Panel::Returncodes>() != UI::Panel::Returncodes::kOk) {
-			return;
-		}
-
-		if (selected_map.empty()) {
-			assert(tree_parent);
-			auto it = tree_parent->subdirectories.find(select.back());
-			assert(it != tree_parent->subdirectories.end());
-			tree_parent->subdirectories.erase(it);
-		} else {
-			auto it = tree->maps.find(selected_map);
-			assert(it != tree->maps.end());
-			tree->maps.erase(it);
-		}
-		break;
-	}
-	}
-
-	addons_with_changes_[m->get_internal_name()] = false;
-	check_for_unsaved_changes();
-	rebuild_dirstruct(m, select);
 }
 
 void AddOnsPackager::die() {
@@ -666,7 +425,7 @@ void AddOnsPackager::clicked_discard_changes() {
 		msg = (boost::format(_("%1$s\n· %2$s")) % msg % str.first).str();
 	}
 
-	UI::WLMessageBox m(get_parent(), UI::WindowStyle::kFsMenu, _("Discard Changes"), msg,
+	UI::WLMessageBox m(&main_menu_, UI::WindowStyle::kFsMenu, _("Discard Changes"), msg,
 	                   UI::WLMessageBox::MBoxType::kOkCancel, UI::Align::kLeft);
 	if (m.run<UI::Panel::Returncodes>() == UI::Panel::Returncodes::kOk) {
 		initialize_mutable_addons();
@@ -687,7 +446,7 @@ void AddOnsPackager::clicked_write_changes() {
 		msg = (boost::format(_("%1$s\n· %2$s")) % msg % str.first).str();
 	}
 
-	UI::WLMessageBox m(get_parent(), UI::WindowStyle::kFsMenu, _("Confirm Saving"), msg,
+	UI::WLMessageBox m(&main_menu_, UI::WindowStyle::kFsMenu, _("Confirm Saving"), msg,
 	                   UI::WLMessageBox::MBoxType::kOkCancel, UI::Align::kLeft);
 	if (m.run<UI::Panel::Returncodes>() == UI::Panel::Returncodes::kOk) {
 		std::set<std::string> errors;
@@ -719,7 +478,7 @@ void AddOnsPackager::clicked_write_changes() {
 }
 
 bool AddOnsPackager::do_write_addon_to_disk(const std::string& addon) {
-	AddOns::Addon* m = mutable_addons_.at(addon);
+	AddOns::Addon* m = mutable_addons_.at(addon).get();
 
 	// Check that the version string is valid and beautify it
 	try {
@@ -740,7 +499,7 @@ bool AddOnsPackager::do_write_addon_to_disk(const std::string& addon) {
 
 	try {
 		return m->write_to_disk();
-	} catch (const WException& e) {
+	} catch (const WLWarning& e) {
 		main_menu_.show_messagebox(
 		   _("Error Writing Addon"),
 		   (boost::format(_("The add-on ‘%1$s’ can not be saved to disk:\n%2$s")) % addon % e.what())
